+++ sdk/dotnet/GetSr.cs
@@ -36,7 +36,7 @@ namespace Pulumi.Xenorchestra
         ///             {
         ///                 SrId = localStorage.Apply(getXoaStorageRepositoryResult =&gt; getXoaStorageRepositoryResult.Id),
         ///                 NameLabel = "Ubuntu Bionic Beaver 18.04_imavo",
-        ///                 Size = "32212254720",
+        ///                 Size = 32212254720,
         ///             },
         ///         },
         ///     });
@@ -71,7 +71,7 @@ namespace Pulumi.Xenorchestra
         ///             {
         ///                 SrId = localStorage.Apply(getXoaStorageRepositoryResult =&gt; getXoaStorageRepositoryResult.Id),
         ///                 NameLabel = "Ubuntu Bionic Beaver 18.04_imavo",
-        ///                 Size = "32212254720",
+        ///                 Size = 32212254720,
         ///             },
         ///         },
         ///     });
@@ -106,7 +106,7 @@ namespace Pulumi.Xenorchestra
         ///             {
         ///                 SrId = localStorage.Apply(getXoaStorageRepositoryResult =&gt; getXoaStorageRepositoryResult.Id),
         ///                 NameLabel = "Ubuntu Bionic Beaver 18.04_imavo",
-        ///                 Size = "32212254720",
+        ///                 Size = 32212254720,
         ///             },
         ///         },
         ///     });
@@ -202,7 +202,7 @@ namespace Pulumi.Xenorchestra
         /// <summary>
         /// The physical storage size.
         /// </summary>
-        public readonly int PhysicalUsage;
+        public readonly double PhysicalUsage;
         /// <summary>
         /// The Id of the pool the storage repository exists on.
         /// </summary>
@@ -210,7 +210,7 @@ namespace Pulumi.Xenorchestra
         /// <summary>
         /// The storage size.
         /// </summary>
-        public readonly int Size;
+        public readonly double Size;
         /// <summary>
         /// The type of storage repository (lvm, udev, iso, user, etc).
         /// </summary>
@@ -222,7 +222,7 @@ namespace Pulumi.Xenorchestra
         /// <summary>
         /// The current usage for this storage repository.
         /// </summary>
-        public readonly int Usage;
+        public readonly double Usage;
         /// <summary>
         /// uuid of the storage repository. This is equivalent to the id.
         /// </summary>
@@ -236,17 +236,17 @@ namespace Pulumi.Xenorchestra
 
             string nameLabel,
 
-            int physicalUsage,
+            double physicalUsage,
 
             string? poolId,
 
-            int size,
+            double size,
 
             string srType,
 
             ImmutableArray<string> tags,
 
-            int usage,
+            double usage,
 
             string uuid)
         {
